Code cleanup; Added storage container check
