+++ AdvancedCrafting.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.2.32630.192
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AdvancedCrafting", "AdvancedCrafting/AdvancedCrafting.csproj", "{5A462C48-3882-4633-8AA1-68848794CF7C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5A462C48-3882-4633-8AA1-68848794CF7C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5A462C48-3882-4633-8AA1-68848794CF7C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5A462C48-3882-4633-8AA1-68848794CF7C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5A462C48-3882-4633-8AA1-68848794CF7C}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {9D37CEB0-5D9E-44D4-AB68-37BE1F0A2559}
+	EndGlobalSection
+EndGlobal
