+++ AdvancedCrafting/Other/Recipe/Ingredient.cs
@@ -1,0 +1,19 @@
+﻿namespace AdvancedCrafting.Other.Recipe
+{
+    public class Ingredient
+    {
+        public ushort Item { get; set; }
+
+        public int Amount { get; set; }
+
+        public Ingredient()
+        {
+        }
+
+        public Ingredient(ushort item, int amount)
+        {
+            Item = item;
+            Amount = amount;
+        }
+    }
+}
