--- AdvancedCrating/Extentions/PlayerExtensions.cs
@@ -1,14 +1,0 @@
-﻿using Rocket.API;
-using Rocket.Unturned.Player;
-using SDG.Unturned;
-
-namespace AdvancedCrating.Extentions
-{
-    public static class PlayerExtensions
-    {
-        public static Player Player(this IRocketPlayer rocketPlayer)
-        {
-            return ((UnturnedPlayer)rocketPlayer).Player;
-        }
-    }
-}
